# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "API CorpTeams",
      "version": "1.0.0"
    },
    "paths": {
      "/funcionarios": {
        "post": {
          "summary": "Cria um novo funcionário",
          "description": "Cria um novo funcionário com os dados fornecidos.",
          "parameters": [
            {
              "name": "funcionario",
              "in": "body",
              "description": "Dados do funcionário a ser criado",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "departamento": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "nome": {
                        "type": "string"
                      }
                    }
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "id": {
                    "type": "integer"
                  },
                  "nome": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Funcionário criado com sucesso"
            },
            "400": {
              "description": "Requisição inválida"
            }
          }
        },
        "get": {
          "summary": "Obtém a lista de funcionários",
          "description": "Retorna a lista de todos os funcionários cadastrados.",
          "responses": {
            "200": {
              "description": "Lista de funcionários obtida com sucesso",
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "departamento": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "nome": {
                          "type": "string"
                        }
                      }
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "id": {
                      "type": "integer"
                    },
                    "nome": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "put": {
          "summary": "Atualiza um funcionário existente",
          "description": "Atualiza os dados de um funcionário existente com base no ID fornecido.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID do funcionário a ser atualizado",
              "required": true,
              "type": "integer"
            },
            {
              "name": "funcionario",
              "in": "body",
              "description": "Novos dados do funcionário",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "departamento": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "nome": {
                        "type": "string"
                      }
                    }
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "nome": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Funcionário atualizado com sucesso"
            },
            "400": {
              "description": "Requisição inválida"
            }
          }
        },
        "delete": {
          "summary": "Remove um funcionário existente",
          "description": "Remove um funcionário com base no ID fornecido.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID do funcionário a ser removido",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "204": {
              "description": "Funcionário removido com sucesso"
            },
            "400": {
              "description": "Requisição inválida"
            }
          }
        }
      }
    }
  }